refactor: display cached color in common charts
